refactor: centralize test utilities
